+++ NOTED/Plugin.cs
@@ -9,9 +9,12 @@ using Dalamud.Interface;
 using Dalamud.Interface.Windowing;
 using Dalamud.Plugin;
 using ImGuiNET;
+using Lumina.Data.Parsing.Uld;
+using NOTED.Helpers;
 using NOTED.Models;
 using NOTED.Windows;
 using System;
+using System.Diagnostics.Metrics;
 using System.Reflection;
 
 namespace NOTED
@@ -79,6 +82,7 @@ namespace NOTED
 
             UiBuilder.Draw += Draw;
             UiBuilder.OpenConfigUi += OpenConfigUi;
+            Framework.Update += Update;
 
             CommandManager.AddHandler(
                 "/noted",
@@ -101,6 +105,10 @@ namespace NOTED
             Settings = Settings.Load();
 
             CreateWindows();
+
+            KeyboardHelper.Initialize();
+
+            SetupKeybinds();
         }
 
         public void Dispose()
@@ -156,6 +164,69 @@ namespace NOTED
                     _settingsWindow.NeedsFocus = true;
                     _settingsWindow.IsOpen = true;
                 }
+            }
+        }
+
+        private void SetupKeybinds()
+        {
+            Settings.HideKeybind.SetAction(() =>
+            {
+                Settings.Hidden = !Settings.Hidden;
+            });
+
+            Settings.NextNoteKeybind.SetAction(() =>
+            {
+                NextNote();
+            });
+
+            Settings.PreviousNoteKeybind.SetAction(() =>
+            {
+                PreviousNote();
+            });
+        }
+
+        private void NextNote()
+        {
+            if (_activeDuty == null || _activeDuty.Notes.Count <= 1) { return; }
+
+            int count = 0;
+
+            do
+            {
+                Note firstNote = _activeDuty.Notes[0];
+                _activeDuty.Notes.Remove(firstNote);
+                _activeDuty.Notes.Add(firstNote);
+                count++;
+            }
+            while (count < _activeDuty.Notes.Count && !_activeDuty.Notes[0].Enabled);
+        }
+
+        private void PreviousNote()
+        {
+            if (_activeDuty == null || _activeDuty.Notes.Count <= 1) { return; }
+
+            int count = 0;
+
+            do
+            {
+                Note lastNote = _activeDuty.Notes[_activeDuty.Notes.Count - 1];
+                _activeDuty.Notes.Remove(lastNote);
+                _activeDuty.Notes.Insert(0, lastNote);
+                count++;
+            }
+            while (count < _activeDuty.Notes.Count && !_activeDuty.Notes[0].Enabled);
+        }
+
+        private void Update(Framework framework)
+        {
+            if (Settings == null || ClientState.LocalPlayer == null) return;
+
+            KeyboardHelper.Instance?.Update();
+
+            KeyBind[] keybinds = Settings.GetKeybinds();
+            foreach (KeyBind keybind in keybinds)
+            {
+                keybind.Update();
             }
         }
 
@@ -227,11 +298,14 @@ namespace NOTED
 
             Settings.Save(Settings);
 
+            KeyboardHelper.Instance?.Dispose();
+
             _windowSystem.RemoveAllWindows();
 
             CommandManager.RemoveHandler("/noted");
             CommandManager.RemoveHandler("/noted toggle");
 
+            Framework.Update -= Update;
             UiBuilder.Draw -= Draw;
             UiBuilder.OpenConfigUi -= OpenConfigUi;
         }