+++ NOTED/changelog.md
@@ -1,5 +1,7 @@
 # 1.1.0.0
 - Added some visibility settings.
+- Added keybind to toggle notes on/off.
+- Added keybinds to cycle through notes.
 
 # 1.0.0.0
 - Added support for Patch 6.3 and Dalamud Api8.